Return ErrEmptyKey if empty key/path is passed to Get and Delete

+++ storage/boltdb/client.go
@@ -116,6 +116,7 @@ func (client *Client) Put(key storage.Key, value storage.Value) error {
 	if key.IsZero() {
 		return storage.ErrEmptyKey.New("")
 	}
+
 	return client.update(func(bucket *bolt.Bucket) error {
 		return bucket.Put(key, value)
 	})
@@ -123,6 +124,10 @@ func (client *Client) Put(key storage.Key, value storage.Value) error {
 
 // Get looks up the provided key from boltdb returning either an error or the result.
 func (client *Client) Get(key storage.Key) (storage.Value, error) {
+	if key.IsZero() {
+		return nil, storage.ErrEmptyKey.New("")
+	}
+
 	var value storage.Value
 	err := client.view(func(bucket *bolt.Bucket) error {
 		data := bucket.Get([]byte(key))
@@ -137,6 +142,10 @@ func (client *Client) Get(key storage.Key) (storage.Value, error) {
 
 // Delete deletes a key/value pair from boltdb, for a given the key
 func (client *Client) Delete(key storage.Key) error {
+	if key.IsZero() {
+		return storage.ErrEmptyKey.New("")
+	}
+
 	return client.update(func(bucket *bolt.Bucket) error {
 		return bucket.Delete(key)
 	})
